requirements.txt
streamlit==1.32.0
tensorflow==2.16.2
numpy==1.26.4
pillow==10.2.0
opencv-python-headless==4.9.0.80